Add changes. Create resonators and waveguides directories
