+++ symplyphysics/laws/electricity/waveguides/attenuation_coefficient_in_dielectric.py
