+++ symplyphysics/laws/electricity/waveguides/attenuation_coefficient_in_dielectric_in_rectangular_waveguide.py
