+++ symplyphysics/laws/electricity/waveguides/attenuation_coefficient_in_metal_in_rectangular_waveguide_for_transverse_magnetic_waves.py
