+++ symplyphysics/laws/electricity/waveguides/characteristic_resistance_of_rectangular_waveguide_for_transverse_electric_waves.py
