+++ symplyphysics/laws/electricity/waveguides/critical_wavelength_in_rectangular_waveguide.py
