+++ symplyphysics/laws/electricity/waveguides/group_velocity_of_wave_in_rectangular_waveguide.py
