+++ symplyphysics/laws/electricity/waveguides/phase_velocity_of_wave_in_rectangular_waveguide.py
