+++ symplyphysics/laws/electricity/waveguides/specific_capacitance_of_coaxial_waveguide.py
