+++ symplyphysics/laws/electricity/waveguides/specific_conductivity_of_coaxial_waveguide.py
