+++ symplyphysics/laws/electricity/waveguides/specific_resistance_of_coaxial_waveguide.py
