+++ symplyphysics/laws/electricity/waveguides/wave_resistance_of_coaxial_waveguide.py
