+++ symplyphysics/laws/electricity/waveguides/wavelength_in_rectangular_waveguide.py
