+++ test/electricity/waveguides/characteristic_resistance_of_rectangular_waveguide_for_transverse_electric_waves_test.py
@@ -2,10 +2,10 @@ from collections import namedtuple
 from pytest import fixture, raises
 from symplyphysics import (errors, units, Quantity, assert_equal,)
 
-from symplyphysics.laws.electricity.circuits import characteristic_resistance_of_rectangular_waveguide_for_transverse_magnetic_waves as resistance_law
+from symplyphysics.laws.electricity.waveguides import characteristic_resistance_of_rectangular_waveguide_for_transverse_electric_waves as resistance_law
 
 ## The critical wavelength is 17.9 millimeters.The wavelength is 10 millimeters. The characteristic resistance
-## of the medium is 254.167 ohms. Then the resistance in the waveguide will be 210.8 ohms.
+## of the medium is 254.167 ohms. Then the resistance in the waveguide will be 306.45 ohms.
 ## https://old.study.urfu.ru/view/aid/67/1/resonators.pdf
 
 Args = namedtuple("Args", ["resistance_of_medium", "wavelength", "critical_wavelength"])
@@ -22,7 +22,7 @@ def test_args_fixture() -> Args:
 def test_basic_resistance(test_args: Args) -> None:
     result = resistance_law.calculate_resistance(test_args.resistance_of_medium, test_args.wavelength,
         test_args.critical_wavelength)
-    assert_equal(result, 210.8 * units.ohm)
+    assert_equal(result, 306.45 * units.ohm)
 
 
 def test_bad_resistance_of_medium(test_args: Args) -> None:
